refactor: move type to types
